refactor: Add sender field to createRoom mutation
This commit modifies the createRoom mutation in the server schema and the APICreateRoom function in RoomUtil.js to include a new field called "sender". This field is used to specify the sender of the room creation request. The sender is added to the list of users in the room and is also set as the sender of the first message in the room.